third_party/WebKit/LayoutTests/fast/css/getComputedStyle/font-family-fallback-reset-expected.txt

Setting a new font-family should reset the fallback list to empty before adding the new family.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(outerDiv, null).fontFamily is "\"courier new\", cursive"
PASS window.getComputedStyle(timesDiv, null).fontFamily is "foo"
PASS window.getComputedStyle(cursiveDiv, null).fontFamily is "cursive"
PASS successfullyParsed is true

TEST COMPLETE
